Scaling and Centering Numerical Data
Improving the model by standardising the numerical data. This ensures the amounts don't influence the models training more than the normalised data from strings. Also adapted the input features. Now take one number (the amount in or out) and the transaction type to tell us where it is in or out. This is more friendly with the standardisation process and the transaction types are instructive to the model.